+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 2.30.0 (Unreleased)
+
+FEATURES:
+
+* **New Data Source:** `aws_acmpca_private_certificate` ([#10183](https://github.com/terraform-providers/terraform-provider-aws/issues/10183))
+* **New Resource:** `aws_acmpca_private_certificate` ([#10183](https://github.com/terraform-providers/terraform-provider-aws/issues/10183))
+*
 ## 2.29.0 (September 20, 2019)
 
 ENHANCEMENTS: